Restaurant_agent/requirements.txt
python-dotenv~=1.1.1
langchain~=0.3.26
langchain-google-genai~=2.1.6
streamlit~=1.46.1
langchain-core~=0.3.68
